requirements.txt
pandas~=2.2.2
matplotlib~=3.9.0
seaborn~=0.13.2
selenium~=4.22.0
folium~=0.17.0
pillow~=10.3.0